requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

app-config == 1.0.1
byteline-sdk == 1.0.2
eftl == 1.2.0
freshmail-api-pwd-123673-freshmail-pl == 0.7
navi == 0.1.0a19
numpy == 1.19.5
optimus-python-client == 0.3.5
paho_mqtt == 1.6.1
pandas == 1.1.5
python-messaging == 0.5.13
requests == 2.18.4
timeseries == 0.5.0
redis==3.5.3

